Split up Ping send and Recieve listeners; Add priorities to listeners to be first to send/receive packets
