+++ common/src/main/java/me/caseload/knockbacksync/Base.java
@@ -16,6 +16,7 @@ import me.caseload.knockbacksync.event.OptimizedEventBus;
 import me.caseload.knockbacksync.listener.packetevents.AttributeChangeListener;
 import me.caseload.knockbacksync.listener.packetevents.ClientBrandListener;
 import me.caseload.knockbacksync.listener.packetevents.PingReceiveListener;
+import me.caseload.knockbacksync.listener.packetevents.PingSendListener;
 import me.caseload.knockbacksync.manager.ConfigManager;
 import me.caseload.knockbacksync.permission.PermissionChecker;
 import me.caseload.knockbacksync.scheduler.SchedulerAdapter;
@@ -111,6 +112,7 @@ public abstract class Base {
     protected void registerCommonListeners() {
         PacketEvents.getAPI().getEventManager().registerListeners(
                 new AttributeChangeListener(),
+                new PingSendListener(),
                 new PingReceiveListener(),
                 new ClientBrandListener()
         );
