+++ common/src/main/java/me/caseload/knockbacksync/listener/packetevents/PingReceiveListener.java
@@ -1,8 +1,6 @@
 package me.caseload.knockbacksync.listener.packetevents;
 
-import com.github.retrooper.packetevents.event.PacketListenerAbstract;
-import com.github.retrooper.packetevents.event.PacketReceiveEvent;
-import com.github.retrooper.packetevents.event.PacketSendEvent;
+import com.github.retrooper.packetevents.event.*;
 import com.github.retrooper.packetevents.protocol.packettype.PacketType;
 import com.github.retrooper.packetevents.protocol.packettype.PacketTypeCommon;
 import com.github.retrooper.packetevents.wrapper.play.client.WrapperPlayClientKeepAlive;
@@ -21,36 +19,9 @@ import java.util.Queue;
 
 public class PingReceiveListener extends PacketListenerAbstract {
 
-    @Override
-    public void onPacketSend(PacketSendEvent event) {
-        if (!Base.INSTANCE.getConfigManager().isToggled()) return;
-
-        PacketTypeCommon packetType = event.getPacketType();
-        if (packetType.equals(PacketType.Play.Server.KEEP_ALIVE)) {
-            WrapperPlayServerKeepAlive keepAlive = new WrapperPlayServerKeepAlive(event);
-            long id = keepAlive.getId();
-
-            PlayerData playerData = PlayerDataManager.getPlayerData(event.getUser().getUUID());
-            if (playerData == null) return;
-
-            playerData.keepaliveMap.add(new Pair<>(id, System.nanoTime()));
-        } else if (packetType.equals(PacketType.Play.Server.PING)) {
-            WrapperPlayServerPing ping = new WrapperPlayServerPing(event);
-            int id = ping.getId();
-
-            PlayerData playerData = PlayerDataManager.getPlayerData(event.getUser().getUUID());
-            if (playerData == null) return;
-
-            playerData.transactionsSent.add(new Pair<>(id, System.nanoTime()));
-        } else if (packetType.equals(PacketType.Play.Server.WINDOW_CONFIRMATION)) {
-            WrapperPlayServerWindowConfirmation confirmation = new WrapperPlayServerWindowConfirmation(event);
-            int id = confirmation.getActionId();
-
-            PlayerData playerData = PlayerDataManager.getPlayerData(event.getUser().getUUID());
-            if (playerData == null) return;
-
-            playerData.transactionsSent.add(new Pair<>(id, System.nanoTime()));
-        }
+    public PingReceiveListener() {
+        // See all incoming ping packets
+        super(PacketListenerPriority.LOWEST);
     }
 
     @Override
